pyproject.toml
[project]
name = "notion_mcp"
version = "0.1.0"
description = "Notion MCP integration for todo lists"
requires-python = ">=3.9"
dependencies = [
    "httpx",
    "python-dotenv"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"